rose metadata-check: remove gtk dependence.
This removes the gtk library dependence from
rose metadata-check and the test-battery tests
for it. The importing of widgets and macros
is now via a function in rose.resource.
