+++ lib/python/rose/config_editor/variable.py
@@ -39,6 +39,7 @@ import rose.config_editor.util
 import rose.gtk.dialog
 import rose.gtk.util
 import rose.reporter
+import rose.resource
 
 
 class VariableWidget(object):
@@ -165,9 +166,9 @@ class VariableWidget(object):
             widget_fpath = os.path.join(lib, *widget_path.split(".")[:-1])
             error_handler = lambda e: self.handle_bad_valuewidget(
                                            str(e), variable, set_value)
-            widget = rose.config_editor.util.import_object(widget_path,
-                                                           files,
-                                                           error_handler)
+            widget = rose.resource.import_object(widget_path,
+                                                 files,
+                                                 error_handler)
             if widget is None:
                 text = rose.config_editor.ERROR_IMPORT_CLASS.format(
                                                                w_val)
